Extract question feature into QuestionPlugin.ts
Editor.tsx had grown to ~1700 lines with a large portion dedicated
entirely to the question feature. Following the existing pattern of
OtherCursorPlugin.ts and QuestionHighlightPlugin.ts, extract all
question-specific code into its own module.

Moved to QuestionPlugin.ts:
- QuestionMapItem interface and Dictionary type
- newQuestionNode / newQuestionAnswerNode factory functions
- QuestionsView, QuestionView, QuestionAnswerView node views
- questionReferenceMarkView and referenceToMarkView mark views
- zIndices/zIndexCounter state and questionPopup function
- questionMarkPlugin factory (with questionMarkWidget helper)
- addQuestion, deleteQuestionSelection, deleteAnswerSelection,
  removeQuestion commands

Editor.tsx retains all wiring: plugin registration, nodeViews/markViews
entries, questionMap property, and addQuestionCommand method.
